app/backend/monitor-demo/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
requests==2.32.3
timeout-decorator==0.5.0
prometheus-client==0.21.1
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-instrumentation-flask==0.41b0
opentelemetry-exporter-prometheus==0.41b0
opentelemetry-instrumentation==0.41b0